[lsp-hover] Fix frame position when it was invisible
The function frame-geometry returns a wrong value for the member
outer-size if the frame is invisible.
The value correspond to the outer-size of the frame the last time it
was visible.
This is probably a bug in Emacs

We now use the functions frame-pixel-{width,height} which return good
values.