lsp-ui-doc: Do not return child frame if it's deleted.
Since the window in the child frame is a dedicated window, it can be killed when
the buffer is killed, resulting in the child frame being deleted. When this
happens, lsp-ui-doc will prevent users from switching to other windows because
the advice on `select-window' fails on operating on the dead frame.

Since there can be other edge cases that may delete the child frame without
notifying lsp-ui-doc, I feel that the most robust solution is always check the
liveness of the child frame before returning it to callers.