Fix same-window logic in select-window advice for doc frame (#250)
* lsp-ui-doc: Truly hide the doc view if selected window changes

Now is an 'around advice to ensure we know both the window that was
selected before and after. Now, it checks that the newly-selected
window is not the previously-selected buffer or the minibuffer.

* Ensure we don't hide the doc frame if it was last selected

Somehow, the frame creation selects the doc-frame's window, so
whenever the doc frame popped up, it would get immediately hidden. We
counteract that by chceking whether the last selected window contained
the doc buffer.