Fix sideline concurrency issues. (#383)
First, currently `lsp-ui-sideline` expects `textDocument/hover` responses to be received in the same order it sends requests. If they are received in different order `lsp-ui-sideline` will delete some of them when received response with index 0.

Second, `lsp-ui-sideline` doesn't cancel old requests that are not needed anymore when user moves the cursor.

Both of these issues are fixed.