[lsp-ui-sideline] Reduce calls to line-number-at-pos
`line-number-at-pos` is the slowest operation in lsp-ui-sideline and was
call several times, including in post-command-hook.
It impacted performance only with big buffers.

This reduce the number of calls to this function to 0 or 1, after the
timer `lsp-ui-sideline-delay`.
It is now called only when the cursor changes of line, when the cursor
stays on the same line it doesn't call `line-number-at-pos`