Also bind "<tab>" in addition to "TAB" for lsp-ui-peek
When "<tab>" (which is the same as `[tab]` I believe) is not bound in `lsp-ui-peek-mode-map`, other keymaps that bind "<tab>" can grab a Tab-key press before it is handled as "TAB" by lsp-ui-peek. Therefore, also bind to "<tab>"/`[tab]` so that the key is always handled by lsp-ui-peek when the peek-window is open. This should only make a difference in GUI emacs, as console emacs can/does not differentiate between "<tab>" and "TAB".