Fix parameter to lsp--make-reference-params
The definition of that function was changed in
emacs-lsp/lsp-mode@2c6a0e92 to reverse the meaning of the parameter,
so here we have to reverse the value we pass in.

This restores the previous behavior of not including declarations in
lsp-ui-peek-find-references by default.

For lsp-ui-find-{next,prev}-reference, reverse the parameters instead,
assuming that when cycling through references, the definition should
be included.